finish moving cancellation activity to TS
Fix #3
