+++ cancellation/src/interfaces/workflows.ts
@@ -1,0 +1,5 @@
+import { Workflow } from '@temporalio/workflow';
+
+export interface Example extends Workflow {
+  main(): Promise<{ message: string }>;
+}
